recalculate contract ltv
