+++ packages/server/customer-os-common-module/service/contract_service.go
@@ -16,12 +16,12 @@ import (
 	neo4jmapper "github.com/openline-ai/openline-customer-os/packages/server/customer-os-neo4j-repository/mapper"
 	neo4jmodel "github.com/openline-ai/openline-customer-os/packages/server/customer-os-neo4j-repository/model"
 	commonpb "github.com/openline-ai/openline-customer-os/packages/server/events-processing-proto/gen/proto/go/api/grpc/v1/common"
-	contractpb "github.com/openline-ai/openline-customer-os/packages/server/events-processing-proto/gen/proto/go/api/grpc/v1/contract"
 	opportunitypb "github.com/openline-ai/openline-customer-os/packages/server/events-processing-proto/gen/proto/go/api/grpc/v1/opportunity"
 	organizationpb "github.com/openline-ai/openline-customer-os/packages/server/events-processing-proto/gen/proto/go/api/grpc/v1/organization"
 	"github.com/opentracing/opentracing-go"
 	"github.com/opentracing/opentracing-go/log"
 	"github.com/pkg/errors"
+	"math"
 	"time"
 )
 
@@ -29,6 +29,7 @@ type ContractService interface {
 	GetById(ctx context.Context, contactId string) (*neo4jentity.ContractEntity, error)
 	Save(ctx context.Context, contactId *string, dataFields data_fields.ContractSaveFields) (string, error)
 	RefreshContractStatus(ctx context.Context, contractId string) error
+	RecalculateContractLtv(ctx context.Context, contractId string) error
 }
 
 type contractService struct {
@@ -328,7 +329,10 @@ func (s *contractService) postUpdateContract(ctx context.Context, tenant string,
 		tracing.TraceErr(span, err)
 		s.log.Errorf("error while updating renewal opportunity for contract %s: %s", contractId, err.Error())
 	}
-	s.updateContractLtv(ctx, tenant, contractId)
+	err = s.RecalculateContractLtv(ctx, contractId)
+	if err != nil {
+		tracing.TraceErr(span, err)
+	}
 	return nil
 }
 
@@ -900,27 +904,6 @@ func (s *contractService) updateActiveRenewalOpportunityLikelihood(ctx context.C
 	return nil
 }
 
-func (s *contractService) updateContractLtv(ctx context.Context, tenant, contractId string) {
-	span, ctx := opentracing.StartSpanFromContext(ctx, "ContractService.updateContractLtv")
-	defer span.Finish()
-	span.SetTag(tracing.SpanTagTenant, tenant)
-	span.SetTag(tracing.SpanTagEntityId, contractId)
-
-	// request contract LTV refresh
-	ctx = tracing.InjectSpanContextIntoGrpcMetadata(ctx, span)
-	_, err := utils.CallEventsPlatformGRPCWithRetry[*contractpb.ContractIdGrpcResponse](func() (*contractpb.ContractIdGrpcResponse, error) {
-		return s.services.GrpcClients.ContractClient.RefreshContractLtv(ctx, &contractpb.RefreshContractLtvGrpcRequest{
-			Tenant:    tenant,
-			Id:        contractId,
-			AppSource: common.GetAppSourceFromContext(ctx),
-		})
-	})
-	if err != nil {
-		tracing.TraceErr(span, err)
-		s.log.Errorf("RefreshContractLtv failed: %s", err.Error())
-	}
-}
-
 func renewalLikelihoodForGrpcRequest(renewalLikelihood neo4jenum.RenewalLikelihood) opportunitypb.RenewalLikelihood {
 	switch renewalLikelihood {
 	case neo4jenum.RenewalLikelihoodHigh:
@@ -983,6 +966,7 @@ func monthsUntilContractEnd(start, end time.Time) int {
 func (s *contractService) RefreshContractStatus(ctx context.Context, contractId string) error {
 	span, ctx := opentracing.StartSpanFromContext(ctx, "ContractService.RefreshContractStatus")
 	defer span.Finish()
+	tracing.SetDefaultServiceSpanTags(ctx, span)
 	span.SetTag(tracing.SpanTagEntityId, contractId)
 
 	// validate tenant
@@ -1008,7 +992,11 @@ func (s *contractService) RefreshContractStatus(ctx context.Context, contractId 
 			tracing.TraceErr(span, err)
 			s.log.Errorf("Error while updating organization relationship for contract %s: %s", contractId, err.Error())
 		}
-		s.updateContractLtv(ctx, tenant, contractId)
+		err = s.RecalculateContractLtv(ctx, contractId)
+		if err != nil {
+			tracing.TraceErr(span, err)
+			s.log.Errorf("Error while updating contract %s ltv: %s", contractId, err.Error())
+		}
 
 		contractDbNode, err := s.services.Neo4jRepositories.ContractReadRepository.GetContractById(ctx, tenant, contractId)
 		if err != nil {
@@ -1037,4 +1025,135 @@ func (s *contractService) RefreshContractStatus(ctx context.Context, contractId 
 	}
 
 	return nil
+}
+
+func (s *contractService) RecalculateContractLtv(ctx context.Context, contractId string) error {
+	span, ctx := opentracing.StartSpanFromContext(ctx, "ContractService.RecalculateContractLtv")
+	defer span.Finish()
+	tracing.SetDefaultServiceSpanTags(ctx, span)
+	span.SetTag(tracing.SpanTagEntityId, contractId)
+
+	// validate tenant
+	err := common.ValidateTenant(ctx)
+	if err != nil {
+		tracing.TraceErr(span, err)
+		return err
+	}
+	tenant := common.GetTenantFromContext(ctx)
+
+	contractDbNode, err := s.services.Neo4jRepositories.ContractReadRepository.GetContractById(ctx, tenant, contractId)
+	if err != nil {
+		tracing.TraceErr(span, err)
+		return err
+	}
+	contractEntity := neo4jmapper.MapDbNodeToContractEntity(contractDbNode)
+
+	ltv := 0.0
+	recalculateContractLtv := true
+	if !(contractEntity.ContractStatus == neo4jenum.ContractStatusLive ||
+		contractEntity.ContractStatus == neo4jenum.ContractStatusOutOfContract ||
+		contractEntity.ContractStatus == neo4jenum.ContractStatusEnded) {
+		span.LogFields(log.String("result", fmt.Sprintf("contract status %s is not eligible for LTV calculation", contractEntity.ContractStatus)))
+		recalculateContractLtv = false
+	}
+
+	if recalculateContractLtv {
+		sliDbNodes, err := s.services.Neo4jRepositories.ServiceLineItemReadRepository.GetServiceLineItemsForContract(ctx, tenant, contractId)
+		if err != nil {
+			tracing.TraceErr(span, err)
+			return err
+		}
+		var sliEntities []*neo4jentity.ServiceLineItemEntity
+		for _, sliDbNode := range sliDbNodes {
+			sliEntities = append(sliEntities, neo4jmapper.MapDbNodeToServiceLineItemEntity(sliDbNode))
+		}
+
+		// Calculate LTV
+
+		// Step 1 calculate one times
+		for _, sliEntity := range sliEntities {
+			if sliEntity.IsOneTime() {
+				sliLtv := float64(sliEntity.Quantity) * sliEntity.Price
+				ltv += sliLtv
+				span.LogFields(log.String("result.sli - ltv", fmt.Sprintf("%s - %f", sliEntity.ID, utils.TruncateFloat64(sliLtv, 2))))
+			}
+		}
+
+		defaultEndDate := utils.Today()
+		if contractEntity.IsEnded() && contractEntity.EndedAt != nil {
+			defaultEndDate = *contractEntity.EndedAt
+		}
+		// Step 2 calculate recurring
+		for _, sliEntity := range sliEntities {
+			if sliEntity.IsRecurrent() {
+				endDate := defaultEndDate
+				if sliEntity.EndedAt != nil && sliEntity.EndedAt.Before(defaultEndDate) {
+					endDate = *sliEntity.EndedAt
+				}
+				duration := calculateDuration(sliEntity.StartedAt, endDate, sliEntity.Billed)
+				sliLtv := float64(sliEntity.Quantity) * sliEntity.Price * duration
+				ltv += sliLtv
+				span.LogFields(log.String("result.sli - ltv", fmt.Sprintf("%s - %f", sliEntity.ID, utils.TruncateFloat64(sliLtv, 2))))
+			}
+		}
+	}
+
+	truncatedLtv := utils.TruncateFloat64(ltv, 2)
+	err = s.services.Neo4jRepositories.ContractWriteRepository.SetLtv(ctx, tenant, contractId, truncatedLtv)
+	if err != nil {
+		tracing.TraceErr(span, err)
+		s.log.Errorf("Error while updating contract %s ltv: %s", contractId, err.Error())
+		return err
+	}
+
+	// get organization for contract
+	organizationDbNode, err := s.services.Neo4jRepositories.OrganizationReadRepository.GetOrganizationByContractId(ctx, tenant, contractId)
+	if err != nil {
+		tracing.TraceErr(span, err)
+		s.log.Errorf("Error while getting organization for contract %s: %s", contractId, err.Error())
+		return nil
+	}
+	organizationEntity := neo4jmapper.MapDbNodeToOrganizationEntity(organizationDbNode)
+
+	// request organization ltv refresh
+	if organizationEntity.ID != "" {
+		ctx = tracing.InjectSpanContextIntoGrpcMetadata(ctx, span)
+		_, err = utils.CallEventsPlatformGRPCWithRetry[*organizationpb.OrganizationIdGrpcResponse](func() (*organizationpb.OrganizationIdGrpcResponse, error) {
+			return s.services.GrpcClients.OrganizationClient.RefreshDerivedData(ctx, &organizationpb.RefreshDerivedDataGrpcRequest{
+				Tenant:         tenant,
+				OrganizationId: organizationEntity.ID,
+				AppSource:      common.GetAppSourceFromContext(ctx),
+			})
+		})
+		if err != nil {
+			tracing.TraceErr(span, err)
+		}
+	}
+
+	s.services.RabbitMQService.PublishEventCompleted(ctx, tenant, contractId, model.CONTRACT, utils.NewEventCompletedDetails().WithUpdate())
+
+	return nil
+}
+
+func calculateDuration(startedAt, endedAt time.Time, billed neo4jenum.BilledType) float64 {
+	if startedAt.After(endedAt) {
+		return float64(0)
+	}
+	durationDays := math.Abs(float64(daysBetween(startedAt, endedAt)))
+
+	switch billed {
+	case neo4jenum.BilledTypeMonthly:
+		return durationDays / 30
+	case neo4jenum.BilledTypeQuarterly:
+		return durationDays / 90
+	case neo4jenum.BilledTypeAnnually:
+		return durationDays / 365
+	default:
+		return 0
+	}
+}
+
+func daysBetween(start, end time.Time) int {
+	duration := end.Sub(start)
+	return int(duration.Hours() / 24)
 }
