+++ packages/server/events-processing-platform/domain/contract/event/events.go
@@ -2,7 +2,6 @@ package event
 
 const (
 	ContractUpdateStatusV1              = "V1_CONTRACT_UPDATE_STATUS"
-	ContractRefreshLtvV1                = "V1_CONTRACT_REFRESH_LTV"
 	ContractRolloutRenewalOpportunityV1 = "V1_CONTRACT_ROLLOUT_RENEWAL_OPPORTUNITY"
 	ContractDeleteV1                    = "V1_CONTRACT_DELETE"
 )
