+++ packages/server/events-processing-proto/grpc/v1/contract.proto
@@ -15,13 +15,6 @@ message RolloutRenewalOpportunityOnExpirationGrpcRequest {
   string appSource = 4;
 }
 
-message RefreshContractLtvGrpcRequest {
-  string id = 1;
-  string tenant = 2;
-  string loggedInUserId = 3;
-  string appSource = 4;
-}
-
 // Permanently delete contract request
 message SoftDeleteContractGrpcRequest {
   string id = 1;
@@ -38,6 +31,5 @@ message ContractIdGrpcResponse {
 // Service definition
 service ContractGrpcService {
   rpc RolloutRenewalOpportunityOnExpiration(RolloutRenewalOpportunityOnExpirationGrpcRequest) returns (ContractIdGrpcResponse);
-  rpc RefreshContractLtv(RefreshContractLtvGrpcRequest) returns (ContractIdGrpcResponse);
   rpc SoftDeleteContract(SoftDeleteContractGrpcRequest) returns (google.protobuf.Empty);
 }
